+++ src/Field.m
@@ -51,9 +51,9 @@
 		[self setName:[coder decodeObject]];
 		[self setSqlField:[coder decodeObject]];
 		[coder decodeValueOfObjCType:@encode(bool) at:&visible];
-		[coder decodeValueOfObjCType:@encode(int) at:&width];
-		[coder decodeValueOfObjCType:@encode(int) at:&tag];
-		[coder decodeValueOfObjCType:@encode(int) at:&type];
+		[coder decodeValueOfObjCType:@encode(NSInteger) at:&width];
+		[coder decodeValueOfObjCType:@encode(NSInteger) at:&tag];
+		[coder decodeValueOfObjCType:@encode(NSInteger) at:&type];
 	}
 	return self;
 }
@@ -116,7 +116,7 @@
 /* setWidth
  * Sets the default width of the field in the article list view.
  */
--(void)setWidth:(int)newWidth
+-(void)setWidth:(NSInteger)newWidth
 {
 	width = newWidth;
 }
@@ -164,7 +164,7 @@
 /* width
  * Returns the default width of the field in the article list view.
  */
--(int)width
+-(NSInteger)width
 {
 	return width;
 }
@@ -182,7 +182,7 @@
  */
 -(NSString *)description
 {
-	return [NSString stringWithFormat:@"('%@', displayName='%@', sqlField='%@', tag=%ld, width=%d, visible=%d)", name, displayName, sqlField, (long)tag, width, visible];
+	return [NSString stringWithFormat:@"('%@', displayName='%@', sqlField='%@', tag=%ld, width=%ld, visible=%d)", name, displayName, sqlField, (long)tag, (long)width, visible];
 }
 
 /* encodeWithCoder
@@ -194,9 +194,9 @@
 	[coder encodeObject:name];
 	[coder encodeObject:sqlField];
 	[coder encodeValueOfObjCType:@encode(bool) at:&visible];
-	[coder encodeValueOfObjCType:@encode(int) at:&width];
-	[coder encodeValueOfObjCType:@encode(int) at:&tag];
-	[coder encodeValueOfObjCType:@encode(int) at:&type];
+	[coder encodeValueOfObjCType:@encode(NSInteger) at:&width];
+	[coder encodeValueOfObjCType:@encode(NSInteger) at:&tag];
+	[coder encodeValueOfObjCType:@encode(NSInteger) at:&type];
 }
 
 /* dealloc
