+++ src/NSDate+Vienna.m
@@ -57,7 +57,7 @@ static BOOL threadSafe;
 	// Initializes the date formatters
 	enUSLocale = [[NSLocale alloc] initWithLocaleIdentifier:@"en_US"];
 
-	for (int i=0; i<kNumberOfDateFormatters; i++)
+	for (NSInteger i=0; i<kNumberOfDateFormatters; i++)
 	{
 		dateFormatterArray[i] = [[NSDateFormatter alloc] init];
 		[dateFormatterArray[i] setLocale:enUSLocale];
@@ -102,7 +102,7 @@ static BOOL threadSafe;
 	{
         // test with the date formatters we are aware of
         // exit as soon as we find a match
-        for (int i=0; i<kNumberOfDateFormatters; i++)
+        for (NSInteger i=0; i<kNumberOfDateFormatters; i++)
         {
             date = [dateFormatterArray[i] dateFromString:modifiedDateString];
             if (date != nil)
@@ -116,7 +116,7 @@ static BOOL threadSafe;
         // NSDateFormatter is thread safe on OS X 10.9 and later only
         // so we manage the issue with this lock
         [dateFormatters_lock lock];
-        for (int i=0; i<kNumberOfDateFormatters; i++)
+        for (NSInteger i=0; i<kNumberOfDateFormatters; i++)
         {
             date = [dateFormatterArray[i] dateFromString:modifiedDateString];
             if (date != nil)
